fix: address ci and copilot findings
