+++ src/styles/scrollbar.test.ts
@@ -1,14 +1,14 @@
-// @ts-nocheck
-
+// @ts-expect-error - app tsconfig omits Node types; Vitest runs this file in Node.
 import { readFileSync } from "node:fs";
+// @ts-expect-error - app tsconfig omits Node types; Vitest runs this file in Node.
 import { fileURLToPath } from "node:url";
+// @ts-expect-error - app tsconfig omits Node types; Vitest runs this file in Node.
 import { dirname, join } from "node:path";
 import { describe, expect, it } from "vitest";
 
 const here = dirname(fileURLToPath(import.meta.url));
 const tokensCss = readFileSync(join(here, "tokens.css"), "utf8");
 const chromeCss = readFileSync(join(here, "chrome.css"), "utf8");
-
 
 function cssRuleBody(css: string, selector: string): string {
   const escaped = selector.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
